+++ physicalai/envs/__init__.py
@@ -1,3 +1,4 @@
 from physicalai.envs.tabletop import TabletopEnv, parse_instruction, ACTIONS, ACTION_NAMES
+from physicalai.envs.gridworld import GridWorld, GW_ACTIONS
 
-__all__ = ["TabletopEnv", "parse_instruction", "ACTIONS", "ACTION_NAMES"]
+__all__ = ["TabletopEnv", "parse_instruction", "ACTIONS", "ACTION_NAMES", "GridWorld", "GW_ACTIONS"]